Extracts shouldShowPrompt function to facilitate code-splitting
