+++ src/index.js
@@ -1,18 +1,7 @@
 import React from "react";
 
+import shouldShowPrompt from "./shouldShowPrompt";
 import PWAPrompt from "./components/PWAPrompt";
-
-const deviceCheck = () => {
-  const isiOS = /iphone|ipad|ipod/.test(
-    window.navigator.userAgent.toLowerCase()
-  );
-  const isiPadOS =
-    navigator.platform === "MacIntel" && navigator.maxTouchPoints > 1;
-  const isStandalone =
-    "standalone" in window.navigator && window.navigator.standalone;
-
-  return (isiOS || isiPadOS) && !isStandalone;
-};
 
 export default ({
   timesToShow = 1,
@@ -27,43 +16,23 @@ export default ({
   debug = false,
   onClose = () => {},
 }) => {
-  let promptData = JSON.parse(localStorage.getItem("iosPwaPrompt"));
+  if (shouldShowPrompt({ timesToShow, promptOnVisit, debug })) {
+    const promptData = JSON.parse(localStorage.getItem("iosPwaPrompt"));
 
-  if (promptData === null) {
-    promptData = { isiOS: deviceCheck(), visits: 0 };
-    localStorage.setItem("iosPwaPrompt", JSON.stringify(promptData));
-  }
-
-  if (promptData.isiOS || debug) {
-    const aboveMinVisits = promptData.visits + 1 >= promptOnVisit;
-    const belowMaxVisits = promptData.visits + 1 < promptOnVisit + timesToShow;
-
-    if (belowMaxVisits || debug) {
-      localStorage.setItem(
-        "iosPwaPrompt",
-        JSON.stringify({
-          ...promptData,
-          visits: promptData.visits + 1,
-        })
-      );
-
-      if (aboveMinVisits || debug) {
-        return (
-          <PWAPrompt
-            delay={delay}
-            copyTitle={copyTitle}
-            copyBody={copyBody}
-            copyAddHomeButtonLabel={copyAddHomeButtonLabel}
-            copyShareButtonLabel={copyShareButtonLabel}
-            copyClosePrompt={copyClosePrompt}
-            permanentlyHideOnDismiss={permanentlyHideOnDismiss}
-            promptData={promptData}
-            maxVisits={timesToShow + promptOnVisit}
-            onClose={onClose}
-          />
-        );
-      }
-    }
+    return (
+      <PWAPrompt
+        delay={delay}
+        copyTitle={copyTitle}
+        copyBody={copyBody}
+        copyAddHomeButtonLabel={copyAddHomeButtonLabel}
+        copyShareButtonLabel={copyShareButtonLabel}
+        copyClosePrompt={copyClosePrompt}
+        permanentlyHideOnDismiss={permanentlyHideOnDismiss}
+        promptData={promptData}
+        maxVisits={timesToShow + promptOnVisit}
+        onClose={onClose}
+      />
+    )
   }
 
   return null;
